pyproject.toml
[project]
name = "oled-spinner"
version = "0.1.0"
description = "Render a gif representation of the oled spinner"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
]
